backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
rdflib==7.0.0
openai==1.3.7
httpx<0.28
pgvector==0.2.4
python-multipart==0.0.6
python-dotenv==1.0.0
redis==5.0.1
